requirements.txt
autopep8==1.5.4
pycodestyle==2.6.0
python-decouple==3.3
selenium==3.141.0
toml==0.10.1
urllib3==1.25.10
